Добавляет разделение кода на модули
